2.3-aiohttp/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==9.0.0
multidict==6.0.4
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==0.21.1
requests==2.28.2
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.40.0
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.12.1
